+++ src/modules/text-module.ts
@@ -563,6 +563,7 @@ export class UltraTextModule extends BaseUltraModule {
       lineHeight: designProperties.line_height || moduleWithDesign.line_height || 'inherit',
       letterSpacing:
         designProperties.letter_spacing || moduleWithDesign.letter_spacing || 'inherit',
+      whiteSpace: designProperties.white_space || moduleWithDesign.white_space || 'inherit',
       margin: '0',
 
       display: 'flex',